requirements.txt
fastapi
pydantic
pyjwt
httpx
uvicorn
sqlmodel
python-multipart
pytest
